requirements.txt
asgiref==3.8.1
boto3==1.34.154
botocore==1.34.154
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==5.0.7
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
